refactor: decompose ItemList into focused row sub-components
Extract the seven internal helpers from ItemList.tsx (633 lines) into
siblings: ItemListRow, ItemListCheckoutRow, ItemListReadOnlyRow,
ItemListCheckoutSubline, ItemListActionMenu, itemListLayout (shared
row geometry constants), and itemListSort. Unify ItemListCheckoutRow
with the read-only variant via an optional onDelete prop. ItemList
narrows to 423 lines as a pure orchestrator.

External API is unchanged — { ItemList } export and prop contract are
preserved; the existing 38-test ItemList suite passes unmodified.
Components use ItemList* prefix to avoid name collision with
features/ai/ItemRow.tsx.
