+++ src/features/bins/itemListLayout.ts
@@ -1,0 +1,6 @@
+// Shared row geometry — keep header, ItemListRow, and ItemListReadOnlyRow aligned.
+export const QTY_COL_WIDTH = 'shrink-0 w-[52px]';
+
+export const ACTIONS_COL_WIDTH = 'shrink-0 w-16';
+
+export const ROW_BASE = 'row-tight px-3.5 py-1 min-h-[44px]';
